Build poor man's static site generator
This is a gross hack and the proper way to do this would be to use
a *real* static site generator (e.g. [react-static](https://github.com/react-static/react-static)).

But until that happens, this *might just be good-enough* to fix the
broken "open link with hash-fragment in new tab" functionality (as
[mentioned on reddit](https://www.reddit.com/r/linux/comments/mkiz74/wayland_explorer_easily_read_wayland_protocol/gtlx9ms/)).