Switch codicons to scoped package name
This is now the preferred way to refer to this package. I've updated to
0.0.21 with this switch, but I haven't gone all the way to 0.0.27 (the
latest as of this writing) because 0.0.22 introduced a breaking change
in the default CSS display property[1]. We could always override it
ourselves, but I don't think the upstream change makes much sense and
I've left a comment to that effect. If they revert the change, it's less
work for us when we do eventually upgrade.

[1] https://github.com/microsoft/vscode-codicons/commit/2afe17f1cbc7a131cf84308e3360f815059cda18