Upgrade react-scripts to v5, tailwindcss to v3, and others
This makes several significant changes to our dependencies:
 - react-scripts (i.e. create-react-app) is now v5 (from v4)
 - tailwindcss is now v3 (from a variant of v2 targeting postcss v7)
 - postcss is now v8 (from v7)
 - autoprefixer is now v10 (from v9)
 - craco is no longer used because react-scripts v5 natively supports
   tailwindcss

The postcss and autoprefixer upgrades required no changes on our end, as
they're used only through tailwindcss.

The craco removal required minor changes to package.json to go back to
invoking react-scripts directly.

The react-scripts upgrade didn't require any changes apart from the
other dependency upgrades. However, it did introduce a pair of harmless
"DeprecationWarning" messages from webpack-dev-server to the build, for
which a fix has already been submitted upstream to create-react-app[1].

The tailwindcss upgrade was the most invasive and required a number of
changes to our CSS and config. I fully followed the upgrade guide[2]
and tested as much functionality as I could. As far as I can tell, the
site behaves and appears identical to how it did before.

[1] https://github.com/facebook/create-react-app/pull/11862
[2] https://tailwindcss.com/docs/upgrade-guide