Add `gamescope` compositor and mention `wlprobe` in new `README`
Gamescope only supports Wayland when ran with `--expose-wayland`,
and only exposes a limited set of protocols from `core` (i.e.
`wl_subcompositor` is not supported, which Rust's `winit` currently
relies on).  This information may be helpful to developers who are
targeting gamescope's Wayland support.

This is generated from version `3.13.17-1` (ArchLinux package).

Also add a README as suggested in [#9] to make it easier for future
readers to know how to (re)generate these files.

[#9]: https://github.com/vially/wayland-explorer/pull/9#issuecomment-1855975715