Remove `linux-dmabuf-unstable-v1`, with redirect to `linux-dmabuf-v1`
This is now a duplicate of the stable `linux-dmabuf-v1`, but (maybe?)
won't be updated in the future for protocol changes. It defines the
exact same interfaces with the exact same names. So there's no reason to
have both.

I think the redirect here should be correct, though I don't know if that
can be tested locally.