Add Jay compositor
Jay is not based on any wayland library (wlroots, weston, smithay, etc.)
and is not represented by any of the other compositors.

You can find a human-readable list of supported protocols [here][here].

Some globals are not exposed to unprivileged clients. When generating
the json file, wlprobe must be invoked as follows: `jay run-privileged
-- wlprobe`.

Jay does not have an icon.

[here]: https://github.com/mahkoh/jay/blob/master/docs/features.md#protocol-support